src/requirements.txt
requests
requests[socks]
pysocks
PyYAML
httpx